Massive code split, each command has now its own file. Common function are now in utils.py
